requirements.txt
streamlit==1.41.1
orange3==3.38.1
numpy==2.1.3
scikit-learn==1.5.2
scipy==1.14.1
pandas==2.2.3
joblib==1.4.2
